fixed indice loading bug; refactored some methods
